requirements.txt
aiosmtplib==3.0.2
aiosqlite==0.21.0
alembic==1.13.3
annotated-types==0.7.0
anyio==4.11.0
APScheduler==3.10.4
asyncpg==0.30.0
bcrypt==5.0.0
blinker==1.9.0
certifi==2025.10.5
cffi==2.0.0
click>=8.1.7,<9.0
colorama==0.4.6
cryptography==46.0.2
Deprecated==1.2.18
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.115.0
fastapi-mail==1.5.0
greenlet==3.2.4
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.27.2
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
limits==5.6.0
loguru==0.7.2
Mako==1.3.10
MarkupSafe==3.0.3
packaging==25.0
passlib==1.7.4
pluggy==1.6.0
psycopg2-binary==2.9.9
pwdlib==0.2.1
pyasn1==0.6.1
pycparser==2.23
pydantic==2.11.10
pydantic-settings==2.11.0
pydantic_core==2.33.2
PyJWT==2.10.1
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-html==4.1.1
python-dotenv==1.1.1
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0.3
redis==5.1.1
rsa==4.9.1
six==1.17.0
slowapi==0.1.9
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.38.6
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.31.0
watchfiles==1.1.0
websockets==15.0.1
win32_setctime==1.2.0
wrapt==1.17.3
